+++ ribbity/render.py
@@ -1,13 +1,28 @@
 from jinja2 import Environment, PackageLoader, FileSystemLoader, ChoiceLoader
 
 
-site_loader = FileSystemLoader("site-templates")
-pkg_loader = PackageLoader("ribbity")
-env = Environment(
-    loader=ChoiceLoader([site_loader, pkg_loader])
-)
+class Piggy:
+    "A container object to be passed into jinja2 render as 'piggy'."
+    def __init__(self, issues_list, labels_to_issues, config_d):
+        self.config = config_d
+        self.issues_list = issues_list
+        self.labels_to_issues = labels_to_issues
 
+        self._init_jinja2()
 
-def render_md(md_name, variables):
-    template = env.get_template(md_name)
-    return template.render(**variables)
+    def _init_jinja2(self):
+        "Initialize jinja2 template loading."
+        site_loader = FileSystemLoader("site-templates")
+        pkg_loader = PackageLoader("ribbity")
+        env = Environment(
+            loader = ChoiceLoader([site_loader, pkg_loader])
+        )
+        self.env = env
+
+    def render(self, template_name, **vardict):
+        "Render given template using 'vardict'. Adds 'piggy' object."
+        render_d = dict(vardict)
+        render_d['piggy'] = self
+
+        template = self.env.get_template(template_name)
+        return template.render(**render_d)
